auth:sanctum

+++ routes/api.php
@@ -3,6 +3,7 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ApiController;
+use App\Http\Controllers\AuthController;
 
 Route::get('ping/', function(){
     return [
@@ -10,8 +11,18 @@ Route::get('ping/', function(){
     ];
 });
 
-Route::post('/todo', [ApiController::class, 'create']);
+Route::get('/unauthenticated', function(){
+    return ['error' => 'Usuário não autenticado!'];
+})->name('login');
+
+Route::post('/user', [AuthController::class, 'create']);
+
+Route::post('/auth', [AuthController::class, 'login']);
+
+Route::middleware('auth:sanctum')->post('/todo', [ApiController::class, 'create']);
+Route::middleware('auth:sanctum')->get('/auth/logout', [AuthController::class, 'logout']);
+
 Route::get('/todos', [ApiController::class, 'all']);
 Route::get('/todo/{id}', [ApiController::class, 'read']);
-Route::put('/todo/{id}', [ApiController::class, 'update']);
-Route::delete('/todo/{id}', [ApiController::class, 'delete']);
+Route::middleware('auth:sanctum')->put('/todo/{id}', [ApiController::class, 'update']);
+Route::middleware('auth:sanctum')->delete('/todo/{id}', [ApiController::class, 'delete']);
